% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{add_env}
\alias{add_env}
\title{Add Environment Variable(s)}
\usage{
add_env(project, user, var, ...)
}
\arguments{
\item{project}{A character string specifying the project name, or an object of class \dQuote{circle_project}. If the latter, there is no need to specify \code{user}.}

\item{user}{A character string specifying the user name. This is not required if \code{project} is of class \dQuote{circle_project}.}

\item{var}{A named list containing key-value pairs of environment variable and value.}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, via \code{\link{circleHTTP}}.}
}
\value{
Something...
}
\description{
Add Circle CI environment variable(s) for a specific project.
}
\details{
Add one or more Circle CI environment variable for a specific project.
}
\examples{
\dontrun{
# add environment variables
add_env(list_projects()[[1]], var = list(A = "123", B = "abc"))

# remove environment variables, one at a time
delete_env(list_projects()[[1]], var = "A")
delete_env(list_projects()[[1]], var = "B")
}
}
\seealso{
\code{\link{delete_env}}
}
