% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/endpoints.R
\name{add_heroku_key}
\alias{add_heroku_key}
\title{Add Heroku key to Circle CI}
\usage{
add_heroku_key(...)
}
\arguments{
\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, via \code{\link{circleHTTP}}.}
}
\value{
Something...
}
\description{
Add Heroku key to Circle CI
}
\details{
Add Heroku key to Circle CI
}

