% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/endpoints.R
\name{authenticate_github}
\alias{authenticate_github}
\title{Add Circle CI key to GitHub}
\usage{
authenticate_github(...)
}
\arguments{
\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, via \code{\link{circleHTTP}}.}
}
\value{
Something...
}
\description{
Add a Circle CI API key to GitHub
}
\details{
Add a Circle CI API key to GitHub.
}

