% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/endpoints.R
\name{cancel_build}
\alias{cancel_build}
\title{Cancel build}
\usage{
cancel_build(build, project, user, ...)
}
\arguments{
\item{build}{A character string specifying the project name, or an object of class \dQuote{circle_build}. If the latter, there is no need to specify \code{project} or \code{user}.}

\item{project}{A character string specifying the project name, or an object of class \dQuote{circle_project}. If the latter, there is no need to specify \code{user}.}

\item{user}{A character string specifying the user name. This is not required if \code{build} is of class \dQuote{circle_build} and/or if \code{project} is of class \dQuote{circle_project} .}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, via \code{\link{circleHTTP}}.}
}
\value{
A list of class \dQuote{circle_build}.
}
\description{
Cancel a Circle CI build.
}
\details{
Cancel a Circle CI build.
}
\examples{
\dontrun{
b <- retry_build(list_builds(limit = 1)[[1]])
cancel_build(b)
}
}
\seealso{
\code{\link{new_build}}, \code{\link{retry_build}}
}

