% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{circleHTTP}
\alias{circleHTTP}
\title{Circle CI HTTP Requests}
\usage{
circleHTTP(verb = "GET", path = "", query = list(), body = "",
  api_version = "v2")
}
\arguments{
\item{verb}{A character string containing an HTTP verb, defaulting to `GET`.}

\item{path}{A character string with the API endpoint (should begin with a
slash).}

\item{query}{A list specifying any query string arguments to pass to the API.
This is used to pass the API token.}

\item{body}{A named list or character string of what should be passed in the
request. Corresponds to the "-d" argument of the `curl` command.}

\item{api_version}{A character string specifying the Circle CI API version.
This usually does not need to be changed by the user.}
}
\value{
The JSON response, or the relevant error.
}
\description{
This is the workhorse function for executing API requests for
  Circle CI.
}
\details{
This is mostly an internal function for executing API requests. In
  almost all cases, users do not need to access this directly.
}
