% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/endpoints.R
\name{delete_cache}
\alias{delete_cache}
\title{Delete Project Cache}
\usage{
delete_cache(project, user, ...)
}
\arguments{
\item{project}{A character string specifying the project name, or an object of class \dQuote{circle_project}. If the latter, there is no need to specify \code{user}.}

\item{user}{A character string specifying the user name. This is not required if \code{build} is of class \dQuote{circle_build} and/or if \code{project} is of class \dQuote{circle_project} .}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, via \code{\link{circleHTTP}}.}
}
\value{
A logical.
}
\description{
Delete project cache
}
\details{
Delete the project cache for a specified Circle CI project.
}
\examples{
\dontrun{
delete_cache(list_projects()[[1]])
}
}

