% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/endpoints.R
\name{delete_env}
\alias{delete_env}
\title{Delete Environment Variable}
\usage{
delete_env(project, user, var, ...)
}
\arguments{
\item{project}{A character string specifying the project name, or an object of class \dQuote{circle_project}. If the latter, there is no need to specify \code{user}.}

\item{user}{A character string specifying the user name. This is not required if \code{project} is of class \dQuote{circle_project}.}

\item{var}{A character string specifying the name of an environment variable}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, via \code{\link{circleHTTP}}.}

\item{branch}{A character string specifying the repository branch.}
}
\value{
A logical.
}
\description{
Delete a Circle CI environment variable for a specific project.
}
\details{
Delete a Circle CI environment variable for a specific project.
}

