% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{enable_project}
\alias{enable_project}
\title{Enable a repo on Circle CI}
\usage{
enable_project(project = NULL, user = NULL, vcs_type = "gh",
  api_version = "v1.1")
}
\arguments{
\item{project}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}

\item{api_version}{A character string specifying the Circle CI API version.
This usually does not need to be changed by the user.}
}
\description{
Follows a repo on Circle CI so that builds can be triggered
}
\examples{
\dontrun{
enable_project()
}
}
