% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builds.R
\name{get_builds}
\alias{get_builds}
\title{Get builds from Circle CI}
\usage{
get_builds(project = NULL, user = NULL, vcs_type = "gh",
  all = FALSE, limit = 30, offset = 30, api_version = "v2")
}
\arguments{
\item{project}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using \code{get_user()}.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}

\item{limit}{How many builds should be returned? Maximum allowed by Circle is
30.}

\item{offset}{}
}
\description{
Queries single builds from Circle CI pipeline runs
}
\details{
To set a different API version, use the following scheme:
\code{https://circleci.com/api/v<api version>} The current default is "v2".
}
