% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_circle_user}
\alias{get_circle_user}
\title{Get Circle CI user}
\usage{
get_circle_user()
}
\value{
A named vector of class \code{circle_user} containing information about
about the authenticated user:
\itemize{
\item Full name
\item Username
\item ID
\item API endpoint
}
}
\description{
Retrieve details about the authenticated Circle CI user.
}
\examples{
\dontrun{
get_circle_user()
}
}
