% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_user}
\alias{get_user}
\title{Get Circle CI user}
\usage{
get_user(...)
}
\arguments{
\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, via \code{\link{circleHTTP}}.}
}
\value{
A list of class \dQuote{circle_user}.
}
\description{
Retrieve details about the authenticated Circle CI user.
}
\details{
This can be used to retrieve your own user profile details and/or as a \dQuote{Hello World!} to test authentication of Circle CI API key specified in \code{Sys.setenv("CIRCLE_CI_KEY" = "exampleapikey")}.
}
\examples{
\dontrun{get_user()}
}

