% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_distance.R
\name{great_circle_distance}
\alias{great_circle_distance}
\title{great_circle_distance}
\usage{
great_circle_distance(lat1, lat2, lon1, lon2, scale = 6371)
}
\arguments{
\item{scale}{}
}
\description{
great_circle_distance
}
\examples{
great_circle_distance(-34, 18, 41, -74)
}
