% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R, R/general.R
\name{list_artifacts}
\alias{list_artifacts}
\title{Get build artifacts}
\usage{
list_artifacts(build = NULL)

list_artifacts(build = NULL)
}
\arguments{
\item{build}{A character string specifying the project name, or an object of class \dQuote{circle_build}. If the latter, there is no need to specify `project` or `user`.}

\item{project}{A character string specifying the project name, or an object of class \dQuote{circle_project}. If the latter, there is no need to specify `user`.}

\item{user}{A character string specifying the user name. This is not required if `build` is of class \dQuote{circle_build} and/or if `project` is of class \dQuote{circle_project} .}

\item{...}{Additional arguments passed to an HTTP request function, such as [httr::GET()], via [circleHTTP()].}

\item{build}{A single Circle CI job object as returned by `get_jobs()`.}
}
\value{
A list of class \dQuote{circle_artifacts}.

A list of build artifacts
}
\description{
Retrieve artifacts from a specific build.

Retrieve artifacts from a specific build.
}
\details{
Retrieves details about artifacts from a specific build.

Retrieves details about artifacts from a specific build.
}
\examples{
\dontrun{
list_artifacts(get_pipelines(limit = 1)[[1]])
}
\dontrun{
list_artifacts(get_builds()[["1"]]$build)
}
}
