% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{retry_build}
\alias{retry_build}
\title{Retry build}
\usage{
retry_build(build = NULL, ...)
}
\arguments{
\item{build}{A character string specifying the project name, or an object of class \dQuote{circle_build}. If the latter, there is no need to specify \code{project} or \code{user}.}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, via \code{\link{circleHTTP}}.}

\item{project}{A character string specifying the project name, or an object of class \dQuote{circle_project}. If the latter, there is no need to specify \code{user}.}

\item{user}{A character string specifying the user name. This is not required if \code{build} is of class \dQuote{circle_build} and/or if \code{project} is of class \dQuote{circle_project} .}
}
\value{
A list of class \dQuote{circle_build}.
}
\description{
Retry a Circle CI build.
}
\details{
Retry a Circle CI build.
}
\examples{
\dontrun{
retry_build(get_pipelines(limit = 1)[[1]])
}
}
\seealso{
\code{\link{cancel_build}}, \code{\link{new_build}}
}
