% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout-key.R
\name{ssh_key}
\alias{ssh_key}
\alias{create_ssh_key}
\alias{get_ssh_keys}
\alias{delete_ssh_key}
\title{Generate/get or delete an SSH Key}
\usage{
create_ssh_key(project = NULL, user = NULL, type = "github-user-key",
  base = "https://circleci.com/api/v1.1", vcs_type = "gh", ...)

get_ssh_keys(project = NULL, user = NULL, vcs_type = "gh", ...)

delete_ssh_key(project = NULL, user = NULL, fingerprint,
  type = "github-user-key", base = "https://circleci.com/api/v1.1",
  vcs_type = "gh", ...)
}
\arguments{
\item{project}{A character string specifying the project name, or an object
of class `circle_project`. If the latter, there is no need to
specify `user`}

\item{user}{A character string specifying the user name. By default the user name
will be queried by `get_user()` and does not need to be set explicitly.}

\item{...}{Additional arguments passed to an HTTP request function, such as
[httr::GET()], via [circleHTTP()].}

\item{key}{A SSH key object as returned by `get_ssh_keys()`}
}
\description{
Generate/get or delete an SSH Key
}
\details{
These functions provide the ability to deal with user SSH keys on
  Circle CI.
}
