% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{test_metadata}
\alias{test_metadata}
\title{Retrieve test metadata}
\usage{
test_metadata(build, project, user, ...)
}
\arguments{
\item{build}{A character string specifying the project name, or an object of class \dQuote{circle_build}. If the latter, there is no need to specify `project` or `user`.}

\item{project}{A character string specifying the project name, or an object of class \dQuote{circle_project}. If the latter, there is no need to specify `user`.}

\item{user}{A character string specifying the user name. This is not required if `build` is of class \dQuote{circle_build} and/or if `project` is of class \dQuote{circle_project} .}

\item{...}{Additional arguments passed to an HTTP request function, such as [httr::GET()], via [circleHTTP()].}
}
\value{
A list of class \dQuote{circle_test_metadata}.
}
\description{
Retrieve test metadata for a build.
}
\details{
Retrieve test metadata for a build.
}
\examples{
\dontrun{
test_metadata(get_pipelines(limit = 1)[[1]])
}
}
