% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{browse_circle_token}
\alias{browse_circle_token}
\title{Authenticate to circle}
\usage{
browse_circle_token()
}
\description{
A circle API Key is needed to interact with the circle API.
  `browse_circle_token()` opens a browser window for the respective circle
  endpoint. On this site, you can copy your personal API key and then follow
  the instructions of the console output or the ones shown below.
}
\section{Store API Key}{


  The `circle` package supports two ways of storing the circle API key(s):

  - via env vars `R_CIRCLE`
  - via `~/.circleci/cli.yml`

  The latter should already be present if you already used the `circle` CLI
  tool at some point in the past. If not, its up to your preference which
  approach to use.

  The following instructions should help to set up `~/.circleci/cli.yml`
  correctly:
  1. Copy the token from the browser window which just opened. You can use
  `edit_circle_config()` to open `~/.circleci/cli.yml`.
  2. The token should be stored using the following structure

     ```
     host: https://circleci.com
     endpoint: graphql-unstable
     token: <token>
     ```
}

