% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builds.R
\name{builds}
\alias{builds}
\alias{get_pipelines}
\alias{get_workflows}
\alias{get_jobs}
\alias{retry_workflow}
\title{Get builds from Circle CI}
\usage{
get_pipelines(
  repo = NULL,
  user = github_info()$owner$login,
  limit = 30,
  vcs_type = "gh",
  api_version = "v2"
)

get_workflows(
  pipeline_id = NULL,
  repo = github_info()$name,
  user = github_info()$owner$login
)

get_jobs(
  workflow_id = NULL,
  repo = github_info()$name,
  user = github_info()$owner$login,
  vcs_type = "gh"
)

retry_workflow(workflow_id = NULL)
}
\arguments{
\item{repo}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{limit}{How many builds should be returned? Maximum allowed by Circle is
30.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}

\item{api_version}{A character string specifying the Circle CI API version.
This usually does not need to be changed by the user.}

\item{pipeline_id}{[string]\cr
A Circle CI pipeline ID.}

\item{workflow_id}{[string]\cr
A Circle CI workflow ID.}
}
\description{
Queries pipelines/workflows/jobs from Circle CI.
}
\examples{
pipelines <- get_pipelines()

workflows <- get_workflows()

jobs <- get_jobs()
}
