% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout-key.R
\name{checkout_key}
\alias{checkout_key}
\alias{create_checkout_key}
\alias{get_checkout_keys}
\alias{delete_checkout_key}
\title{Generate/get or delete an checkout Key}
\usage{
create_checkout_key(project = NULL, user = NULL, type = "deploy-key",
  api_version = "v1.1", vcs_type = "gh")

get_checkout_keys(project = NULL, user = NULL, vcs_type = "gh")

delete_checkout_key(project = NULL, user = NULL, fingerprint,
  type = "github-user-key", api_version = "v1.1", vcs_type = "gh")
}
\arguments{
\item{project}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{type}{Type of key to add. Options are "github-user-key" and
"deploy-key".}

\item{api_version}{A character string specifying the Circle CI API version.
This usually does not need to be changed by the user.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}

\item{fingerprint}{The fingerprint of the checkout key which should be deleted.}
}
\description{
Generate/get or delete an checkout Key
}
\details{
These functions provide the ability to deal with user checkout keys on
  Circle CI.
}
