% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout-key.R
\name{checkout_key}
\alias{checkout_key}
\alias{create_checkout_key}
\alias{get_checkout_keys}
\alias{delete_checkout_key}
\alias{has_checkout_key}
\title{Add, Get or Delete a Checkout Keys on Circle CI}
\usage{
create_checkout_key(
  repo = github_info()$name,
  user = github_info()$owner$login,
  type = "user-key",
  api_version = "v2",
  vcs_type = "gh"
)

get_checkout_keys(
  repo = github_info()$name,
  user = github_info()$owner$login,
  vcs_type = "gh",
  api_version = "v2"
)

delete_checkout_key(
  fingerprint = NULL,
  repo = github_info()$name,
  user = github_info()$owner$login,
  type = "user-key",
  api_version = "v2",
  vcs_type = "gh"
)

has_checkout_key(
  repo = github_info()$name,
  user = github_info()$owner$login,
  type = "github-user-key",
  vcs_type = "gh",
  preferred = TRUE
)
}
\arguments{
\item{repo}{\verb{[character]}\cr
The repository slug to use. Must follow the "\code{user/repo}" structure.}

\item{user}{\verb{[character]}\cr
The username for the repository. By default queried using \code{get_user()}.}

\item{type}{\verb{[character]}\cr
Type of key to add. Options are "user-key" and "deploy-key".}

\item{api_version}{\verb{[character]}\cr
A character string specifying the Circle CI API version.
This usually does not need to be changed by the user.}

\item{vcs_type}{\verb{[character]}\cr The version control system to use.
Defaults to "gh" (Github).}

\item{fingerprint}{\verb{[character]}\cr
The fingerprint of the checkout key which should be deleted.}

\item{preferred}{\verb{[logical]}\cr
Checks whether the requested type is the "preferred" key.}
}
\description{
Functions to interact with "checkout keys" on Circle CI.
}
\examples{
\dontrun{
create_checkout_key()

# create only a 'deploy key' pair
create_checkout_key(type = "deploy-key")
}
\dontrun{
get_checkout_keys()
}
\dontrun{
delete_checkout_key()
}
\dontrun{
has_checkout_key()
}
}
