% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout-key.R
\name{checkout_key}
\alias{checkout_key}
\alias{create_checkout_key}
\alias{get_checkout_keys}
\alias{delete_checkout_key}
\alias{has_checkout_key}
\title{Generate/get or delete an checkout Key}
\usage{
create_checkout_key(
  repo = github_info()$name,
  user = github_info()$owner$login,
  type = "user-key",
  api_version = "v2",
  vcs_type = "gh"
)

get_checkout_keys(
  repo = github_info()$name,
  user = github_info()$owner$login,
  vcs_type = "gh",
  api_version = "v2"
)

delete_checkout_key(
  fingerprint = NULL,
  repo = github_info()$name,
  user = github_info()$owner$login,
  type = "user-key",
  api_version = "v2",
  vcs_type = "gh"
)

has_checkout_key(
  repo = github_info()$name,
  user = github_info()$owner$login,
  type = "github-user-key",
  vcs_type = "gh",
  preferred = "true"
)
}
\arguments{
\item{repo}{The Circle CI project. By default the current directory name will
be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{type}{Type of key to add. Options are "user-key" and
"deploy-key".}

\item{api_version}{A character string specifying the Circle CI API version.
This usually does not need to be changed by the user.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}

\item{fingerprint}{The fingerprint of the checkout key which should be
deleted.}

\item{preferred}{Checks whether the requested type is the "preferred" key.}
}
\description{
Functions to interact with "checkout keys" on Circle CI.

Check if a specific key type exists in the Circle CI project
}
\examples{
\dontrun{
create_checkout_key()

# create only a 'deploy key' pair

create_checkout_key(type = "deploy-key")
}
\dontrun{
get_checkout_keys()
}
\dontrun{
delete_checkout_key()
}
\dontrun{
has_checkout_key()
}
}
