% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{circleHTTP}
\alias{circleHTTP}
\title{Circle CI HTTP Requests}
\usage{
circleHTTP(verb = "GET", path = "", query = list(), body = "",
  api_version = "v2", encode = "json", ...)
}
\arguments{
\item{verb}{A character string containing an HTTP verb, defaulting to \dQuote{GET}.}

\item{path}{A character string with the API endpoint (should begin with a slash).}

\item{query}{A list specifying any query string arguments to pass to the API.}

\item{body}{character string of request body data.}

\item{...}{Additional arguments passed to an HTTP request function, such as [httr::GET()].}

\item{base}{A character string specifying the base URL for the API.}

\item{key}{A character string containing a Circle CI API token. If missing, defaults to value stored in environment variable \dQuote{CIRCLE_CI_KEY}.}
}
\value{
The JSON response, or the relevant error.
}
\description{
This is the workhorse function for executing API requests for Circle CI.
}
\details{
This is mostly an internal function for executing API requests.
In almost all cases, users do not need to access this directly.
}
