% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http.R
\name{circleHTTP}
\alias{circleHTTP}
\title{Circle CI HTTP Requests}
\usage{
circleHTTP(verb = "GET", path = "", query = list(), body = "",
  base = "https://circleci.com/api/v1", key = Sys.getenv("CIRCLE_CI_KEY"),
  ...)
}
\arguments{
\item{verb}{A character string containing an HTTP verb, defaulting to \dQuote{GET}.}

\item{path}{A character string with the API endpoint (should begin with a slash).}

\item{query}{A list specifying any query string arguments to pass to the API.}

\item{body}{character string of request body data.}

\item{key}{A character string containing a Circle CI API token. If missing, defaults to value stored in environment variable \dQuote{CIRCLE_CI_KEY}.}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}.}
}
\value{
The JSON response, or the relevant error.
}
\description{
This is the workhorse function for executing API requests for Circle CI.
}
\details{
This is mostly an internal function for executing API requests.
In almost all cases, users do not need to access this directly.
}

