% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{delete_cache}
\alias{delete_cache}
\title{Delete Project Cache}
\usage{
delete_cache(project = NULL, user = NULL, vcs_type = "gh")
}
\arguments{
\item{project}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}
}
\value{
A logical.
}
\description{
Delete project cache
}
\details{
Delete the project cache for a specified Circle CI project.
}
\examples{
\dontrun{
delete_cache(list_projects()[[1]])
}
}
