% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R, R/general.R
\name{delete_cache}
\alias{delete_cache}
\title{Delete Project Cache}
\usage{
delete_cache(project = NULL, user = NULL, vcs_type = "gh")

delete_cache(project = NULL, user = NULL, vcs_type = "gh")
}
\arguments{
\item{project}{A character string specifying the project name, or an object of class \dQuote{circle_project}. If the latter, there is no need to specify `user`.}

\item{user}{A character string specifying the user name. This is not required if `build` is of class \dQuote{circle_build} and/or if `project` is of class \dQuote{circle_project} .}

\item{...}{Additional arguments passed to an HTTP request function, such as [httr::GET()], via [circleHTTP()].}

\item{project}{A character string specifying the project name, or an object
of class \dQuote{circle_project}. If the latter, there is no need to
specify `user`.}

\item{user}{A character string specifying the user name. This is not required
if `build` is of class \dQuote{circle_build} and/or if `project` is of
class \dQuote{circle_project} .}

\item{...}{Additional arguments passed to an HTTP request function, such as
[httr::GET()], via [circleHTTP()].}
}
\value{
A logical.

A logical.
}
\description{
Delete project cache

Delete project cache
}
\details{
Delete the project cache for a specified Circle CI project.

Delete the project cache for a specified Circle CI project.
}
\examples{
\dontrun{
delete_cache(list_projects()[[1]])
}
\dontrun{
delete_cache(list_projects()[[1]])
}
}
