% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{delete_cache}
\alias{delete_cache}
\title{Delete repo Cache}
\usage{
delete_cache(repo = github_info()$name,
  user = get_user()$content$login, vcs_type = "gh")
}
\arguments{
\item{repo}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}
}
\value{
A logical.
}
\description{
Delete repo cache
}
\details{
Delete the repo cache for a specified Circle CI repo.
}
\examples{
\dontrun{
delete_cache(list_repos()[[1]])
}
}
