% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{edit_circle_config}
\alias{edit_circle_config}
\title{Open circle Configuration file}
\usage{
edit_circle_config()
}
\value{
No return value, called for side effects.
}
\description{
Opens \verb{~/.circleci/cli.yml}.
}
