% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{enable_repo}
\alias{enable_repo}
\title{Enable a repo on Circle CI}
\usage{
enable_repo(
  repo = github_info()$name,
  user = github_info()$owner$login,
  vcs_type = "gh",
  api_version = "v1.1"
)
}
\arguments{
\item{repo}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}

\item{api_version}{A character string specifying the Circle CI API version.
This usually does not need to be changed by the user.}
}
\description{
Follows a repo on Circle CI so that builds can be triggered
}
\examples{
\dontrun{
enable_repo()
}
}
