% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-var.R
\name{env_var}
\alias{env_var}
\alias{get_env_vars}
\alias{set_env_var}
\alias{delete_env_var}
\title{Interact with Environment Variable(s)}
\usage{
get_env_vars(
  name = NULL,
  repo = github_info()$name,
  user = get_user()$content$login,
  vcs_type = "gh",
  api_version = "v2"
)

set_env_var(
  var,
  repo = github_info()$name,
  user = get_user()$content$login,
  vcs_type = "gh",
  api_version = "v2",
  quiet = FALSE
)

delete_env_var(
  var,
  repo = github_info()$name,
  user = get_user()$content$login,
  vcs_type = "gh",
  api_version = "v2",
  quiet = FALSE
)
}
\arguments{
\item{name}{Name of an environment variable. If not set, all env vars are
returned.}

\item{repo}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}

\item{api_version}{A character string specifying the Circle CI API version.
This usually does not need to be changed by the user.}

\item{var}{A list containing key-value pairs of environment variable
and its value.}

\item{quiet}{If `TRUE`, console output is suppressed.}
}
\description{
Add/Get/Set Circle CI environment variable(s) for a repo on
  Circle CI.
}
\examples{
\dontrun{
 # get env var
 get_env_vars()

 # set env var
 set_env_var(var = list("foo" = "123"))

 # delete env var
 delete_env_var("foo")
}
}
