% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{generate_ssh_key}
\alias{generate_ssh_key}
\title{Generate SSH Key}
\usage{
generate_ssh_key(project, user, ...)
}
\arguments{
\item{project}{A character string specifying the project name, or an object of class \dQuote{circle_project}. If the latter, there is no need to specify \code{user}.}

\item{user}{A character string specifying the user name. This is not required if \code{build} is of class \dQuote{circle_build} and/or if \code{project} is of class \dQuote{circle_project} .}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, via \code{\link{circleHTTP}}.}
}
\value{
Something...
}
\description{
Generate an SSH Key
}
\details{
Generate an SSH Key for a specific project
}
