% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_build_artifacts}
\alias{get_build_artifacts}
\title{Get build artifacts}
\usage{
get_build_artifacts(
  job_id = NULL,
  repo = github_info()$name,
  user = github_info()$owner$login,
  vcs_type = "gh",
  api_version = "v2"
)
}
\arguments{
\item{job_id}{A Circle CI job id.}

\item{repo}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}

\item{api_version}{A character string specifying the Circle CI API version.
This usually does not need to be changed by the user.}
}
\value{
A list of build artifacts
}
\description{
Retrieve artifacts from a specific build.
}
\details{
Retrieves details about artifacts from a specific build.
}
\examples{
\dontrun{
list_artifacts(get_builds()[["1"]]$build)
}
}
