% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builds.R
\name{get_pipelines}
\alias{get_pipelines}
\title{Get builds from Circle CI}
\usage{
get_pipelines(
  repo = NULL,
  user = github_info()$owner$login,
  limit = 30,
  vcs_type = "gh",
  api_version = "v2"
)
}
\arguments{
\item{repo}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{limit}{How many builds should be returned? Maximum allowed by Circle is
30.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}

\item{api_version}{A character string specifying the Circle CI API version.
This usually does not need to be changed by the user.}
}
\description{
Queries single builds from Circle CI pipeline runs
}
