% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_user}
\alias{get_user}
\title{Get Circle CI user}
\usage{
get_user()
}
\value{
A list of class `circle_user`.
}
\description{
Retrieve details about the authenticated Circle CI user.
}
\examples{
\dontrun{
get_user()
}
}
