% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_user}
\alias{get_user}
\title{Get Circle CI user}
\usage{
get_user()
}
\value{
A list of class `circle_user`.
}
\description{
Retrieve details about the authenticated Circle CI user.
}
\details{
This can be used to retrieve your own user profile details and/or as
  a `Hello World!` to test authentication of Circle CI API key
  specified in `Sys.setenv("CIRCLE_CI_KEY" = "exampleapikey")`.
}
\examples{
\dontrun{get_user()}
}
