% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-github.R
\name{github_info}
\alias{github_info}
\title{Github information}
\usage{
github_info(path = ".", remote = "origin", .token = NULL)
}
\arguments{
\item{path}{`[string]`\cr
The path to a GitHub-enabled Git repository (or a subdirectory thereof).}

\item{remote}{`[string]`\cr
The Github remote which should be used. Defaults to "origin".}

\item{.token}{`[character]`\cr
Authentication token. Defaults to GITHUB_PAT or GITHUB_TOKEN environment
variables, in this order if any is set. See gh_token() if you need more
flexibility, e.g. different tokens for different GitHub Enterprise
deployments.}
}
\description{
Retrieves metadata about a Git repository from GitHub.

`github_info()` returns a list as obtained from the GET "/repos/:repo" API.
}
\concept{GitHub functions}
