% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-helpers.R
\name{github_info}
\alias{github_info}
\title{Github information}
\usage{
github_info(path = rprojroot::find_package_root_file())
}
\arguments{
\item{path}{`[string]`\cr
The path to a GitHub-enabled Git repository (or a subdirectory thereof).}
}
\description{
Retrieves metadata about a Git repository from GitHub.

`github_info()` returns a list as obtained from the GET "/repos/:repo" API.
}
\concept{GitHub functions}
