% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{list_artifacts}
\alias{list_artifacts}
\title{Get build artifacts}
\usage{
list_artifacts(build = NULL)
}
\arguments{
\item{build}{A single Circle CI job object as returned by `get_jobs()`.}
}
\value{
A list of build artifacts
}
\description{
Retrieve artifacts from a specific build.
}
\details{
Retrieves details about artifacts from a specific build.
}
\examples{
\dontrun{
list_artifacts(get_builds()[["1"]]$build)
}
}
