% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{list_projects}
\alias{list_projects}
\title{List repos}
\usage{
list_projects()
}
\value{
A list of `circle_repo`s.
}
\description{
Retrieve a list of Circle CI repos for the authenticated
  user.
}
\details{
Retrieves a very detailed list of repos and repo-related
  information for all Circle CI repos attached to the current user.
}
\examples{
\dontrun{
repos <- list_repos()
}
}
\seealso{
[get_builds()], [get_pipelines()]
}
