% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{list_projects}
\alias{list_projects}
\title{List projects}
\usage{
list_projects()
}
\value{
A list of `circle_project`s.
}
\description{
Retrieve a list of Circle CI projects for the authenticated
  user.
}
\details{
Retrieves a very detailed list of projects and project-related
  information for all Circle CI projects attached to the current user.
}
\examples{
\dontrun{
projects = list_projects()
}
}
\seealso{
[get_builds()], [get_pipelines()]
}
