% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{list_projects}
\alias{list_projects}
\title{List projects}
\usage{
list_projects(...)
}
\arguments{
\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr:GET]{httr::GET()}}, via \code{\link[=circleHTTP]{circleHTTP()}}.}
}
\value{
A list of class \dQuote{circle_projects}, wherein each element is a \dQuote{circle_project}.
}
\description{
Retrieve a list of Circle CI projects for the authenticated user.
}
\details{
Retrieves a very detailed list of projects and project-related information for all Circle CI projects attached to the current user.
}
\examples{
\dontrun{list_projects()}
}
\seealso{
\code{\link[=get_build]{get_build()}}, \code{\link[=get_pipelines]{get_pipelines()}}
}
