% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{list_projects}
\alias{list_projects}
\title{List repository}
\usage{
list_projects(repo = github_info()$name, user = github_info()$owner$login)
}
\arguments{
\item{repo}{The Circle CI project. By default the current directory name will
be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}
}
\value{
An object of class `circle_api`.
}
\description{
Retrieve a list of Circle CI repository for the authenticated
  user.
}
\details{
Retrieves a very detailed list of repository and repo-related
  information for all Circle CI repository attached to the current user.

  This endpoint uses API v1.1 and will probably be removed in the near
  future.
}
\examples{
\dontrun{
list_projects()
}
}
\seealso{
[get_pipelines()], [get_workflows()], [get_jobs()]
}
