% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{list_projects}
\alias{list_projects}
\title{List repos}
\usage{
list_projects(repo = github_info()$name, user = github_info()$owner$login)
}
\arguments{
\item{repo}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}
}
\value{
A list of `circle_repo`s.
}
\description{
Retrieve a list of Circle CI repos for the authenticated
  user.
}
\details{
Retrieves a very detailed list of repos and repo-related
  information for all Circle CI repos attached to the current user.

  This endpoint uses API v1.1 and will probably be removed in the near
  future.
}
\examples{
\dontrun{
repos <- list_repos()
}
}
\seealso{
[get_builds()], [get_pipelines()]
}
