% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R, R/general.R
\name{list_projects}
\alias{list_projects}
\title{List projects}
\usage{
list_projects()

list_projects()
}
\arguments{
\item{...}{Additional arguments passed to an HTTP request function, such as [httr::GET()], via [circleHTTP()].}
}
\value{
A list of class \dQuote{circle_projects}, wherein each element is a \dQuote{circle_project}.

A list of `circle_project`s.
}
\description{
Retrieve a list of Circle CI projects for the authenticated user.

Retrieve a list of Circle CI projects for the authenticated
  user.
}
\details{
Retrieves a very detailed list of projects and project-related information for all Circle CI projects attached to the current user.

Retrieves a very detailed list of projects and project-related
  information for all Circle CI projects attached to the current user.
}
\examples{
\dontrun{list_projects()}
projects = list_projects()
}
\seealso{
[get_build()], [get_pipelines()]

[get_builds()], [get_pipelines()]
}
