% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R, R/general.R
\name{new_build}
\alias{new_build}
\title{Trigger build}
\usage{
new_build(project = NULL, user = NULL, vcs_type = "gh",
  branch = "master")

new_build(project = NULL, user = NULL, vcs_type = "gh",
  branch = "master")
}
\arguments{
\item{project}{A character string specifying the project name, or an object of class \dQuote{circle_project}. If the latter, there is no need to specify `user`.}

\item{user}{A character string specifying the user name. This is not required if `project` is of class \dQuote{circle_project}.}

\item{branch}{A character string specifying the repository branch.}

\item{...}{Additional arguments passed to an HTTP request function, such as [httr::GET()], via [circleHTTP()].}

\item{project}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{branch}{A character string specifying the repository branch.}
}
\value{
A list of class \dQuote{circle_build}.
}
\description{
Trigger a new build for a specific project branch

Trigger a new build for a specific project branch
}
\details{
Trigger a new Circle CI build for a specific project branch.

Trigger a new Circle CI build for a specific project branch.
}
\examples{
\dontrun{
p <- list_projects()
new_build(p[[1]])
}
\dontrun{
new_build()
}
}
\seealso{
[cancel_build()], [retry_build()]

[retry_build()]
}
