% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{new_build}
\alias{new_build}
\title{Trigger build}
\usage{
new_build(project, user, branch, ...)
}
\arguments{
\item{project}{A character string specifying the project name, or an object of class \dQuote{circle_project}. If the latter, there is no need to specify \code{user}.}

\item{user}{A character string specifying the user name. This is not required if \code{project} is of class \dQuote{circle_project}.}

\item{branch}{A character string specifying the repository branch.}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr:GET]{httr::GET()}}, via \code{\link[=circleHTTP]{circleHTTP()}}.}
}
\value{
A list of class \dQuote{circle_build}.
}
\description{
Trigger a new build for a specific project branch
}
\details{
Trigger a new Circle CI build for a specific project branch.
}
\examples{
\dontrun{
p <- list_projects()
new_build(p[[1]])
}
}
\seealso{
\code{\link[=cancel_build]{cancel_build()}}, \code{\link[=retry_build]{retry_build()}}
}
