% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{new_build}
\alias{new_build}
\title{Trigger build}
\usage{
new_build(repo = NULL, user = NULL, vcs_type = "gh",
  branch = "master")
}
\arguments{
\item{repo}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}

\item{branch}{A character string specifying the repository branch.}
}
\description{
Trigger a new build for a specific repo branch
}
\details{
Trigger a new Circle CI build for a specific repo branch.
}
\examples{
\dontrun{
new_build()
}
}
\seealso{
[retry_build()]
}
