% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout-key.R
\name{ssh_key}
\alias{ssh_key}
\alias{create_ssh_key}
\alias{get_ssh_keys}
\alias{delete_ssh_key}
\title{Generate/get or delete an SSH Key}
\usage{
create_ssh_key(project = NULL, user = NULL, type = "github-user-key",
  api_version = "https://circleci.com/api/v1.1", vcs_type = "gh")

get_ssh_keys(project = NULL, user = NULL, vcs_type = "gh")

delete_ssh_key(project = NULL, user = NULL, fingerprint,
  type = "github-user-key",
  api_version = "https://circleci.com/api/v1.1", vcs_type = "gh")
}
\arguments{
\item{project}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}

\item{type}{Type of key to add. Options are "github-user-key" and
"deploy-key".}

\item{api_version}{A character string specifying the Circle CI API version.
This usually does not need to be changed by the user.}

\item{vcs_type}{The version control system to use. Defaults to "gh" (Github).
Other possible values are "bitbucket".}

\item{fingerprint}{The fingerprint of the SSH key which should be deleted.}
}
\description{
Generate/get or delete an SSH Key
}
\details{
These functions provide the ability to deal with user SSH keys on
  Circle CI.
}
