% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-circle-deploy.R
\name{use_circle_deploy}
\alias{use_circle_deploy}
\title{Set up deployment between Circle CI builds and Github repositories}
\usage{
use_circle_deploy(repo = github_info()$name, user = github_info()$owner$login)
}
\arguments{
\item{repo}{The Circle CI project. By default the current directory name will be used.}

\item{user}{The username for the repository. By default queried using `get_user()`.}
}
\description{
Checks and eventually creates a "github user key" to enable
  deployment from Circle CI builds.
}
\details{
The easiest way to achieve a deployment from Circle CI builds to a Github
repo is by creating a so called "github user key" with your Circle CI
account.

This function checks for the presence of such and creates one, if missing.

If the repo has not been enabled yet on Circle CI, please run `enable_repo()`
first. Also to be able to log in to Github, you will need to have a
`GITHUB_TOKEN` being set as an environment variable. If you are unsure
if you have done this already, call `usethis::github_token()`.
If none is set, this function will prompt you to create one.
}
\examples{
\dontrun{
use_circle_deploy()
}
}
